Added DataWriter
Added DataWriter class as a wrapper to ofstream to write data from a TimeFrequency buffer to file. Has a constructor that takes a string of the file path and one method "write_time_frequency_data_to_file" for writing the whole contents of a TimeFrequency buffer.

Added basic unit tests for DataWriter.

Refactored the name of the read method in DataReader to "read_time_frequency_data_from_file".
Refactored GetAbsoluteFilepathFromRelative into the global namespace.