Added RfiStrategy
Added RfiStrategy class using the CRTP pattern as a framework for selecting the type of RFIM to use from several available options in a way that works with the templated TimeFrequency buffer.
Currently contains a process method which accepts a TimeFrequency.

Added RudimentaryRfi class as a concrete RfiStrategy. This is based on the initial example code given in "rfi_clean.cpp" and has been modified a minimum amount from this to work with this projects architecture.

Also added a very basic test for RudimentaryRfi.