Added FileProcessor
Added FileProcessor template class for applying a chosen RfiStrategy to a file, cleaning it and saving to a new file. process_file will loop over the full number of whole chunks in the input file and returns the number of chunks iterated over.

Updated DataReader to have get_file_length and get_file_length_bytes methods to get the full length of the file. This value is stored as a member variable in the constructor. Also added get_remaining_file_length and get_remaining_file_length_bytes methods. NOTE: if the file is modified while DataReader is reading it, this could lead to errors.

Updated RudimentaryRfi with static_assert for templated type. The DataType is now also captured.

Added new tests for expanded DataReader functionality.
Added a basic test for FileProcessor.