Added MedianStandardDeviationRfi
Added RfiStrategy MedianStandardDeviationRfi. This is based on the method used in "rfi_clean.cpp", but has been optimised for better performance.
As a first pass basic attempt at optimising, it now uses std::nth_element to find the median.
The performance is orders of magnitudes better than the original.

Added a basic test for MedianStandardDeviationRfi that uses the full data.bin