Added rfim_demo
Added: rfim_demo project.
This is a command line executable that demonstrates some basic functionality of the rfim library. The RfiStrategy can be selected and the test file will be cleaned.

Updated: README with info about rfim_demo.

Updated: FileProcessor and FileProcessorInfo  with to measure and record time spent processing data.